// apps/backend/src/app/notifications/notification-type.enum.ts
export enum NotificationType {
  NEW_APPLICATION = 'NEW_APPLICATION',
  APPLICATION_SUBMITTED = 'APPLICATION_SUBMITTED',
  APPLICATION_STATUS_UPDATE = 'APPLICATION_STATUS_UPDATE',
  APPLICATION_REJECTED = 'APPLICATION_REJECTED',
  NEW_JOB = 'NEW_JOB',
  JOB_MATCH = 'JOB_MATCH',
  JOB_RECOMMENDATION = 'JOB_RECOMMENDATION',
  SAVED_JOB_UPDATE = 'SAVED_JOB_UPDATE',
  AI_RESUME_PARSED = 'AI_RESUME_PARSED',
  AI_RESUME_SCORED = 'AI_RESUME_SCORED',
  RECOMMENDATION = 'RECOMMENDATION',
  PERSONALIZED_RECOMMENDATION = 'PERSONALIZED_RECOMMENDATION',
}